New CLIST command / CARCHIVE stores container size
The new CLIST command returns a list of containers inside the status document. For this the core ngamsStatus class was modified so it could hold container references. The CCREATE command is also returning the list of created containers with the Status document that is sent back to the client. The CARCHIVE command should probably do the same.

On a separate topic, the CARCHIVE command now stores the total amount of space that the container will use if all its contents are retrieved. This doesn't consider inner containers' sizes, but that's probably something that can be deal with on the client side if necessary