ngamsContainer contains ngamsFileInfo instances
Now containers refer to their file objects by pointing at them. Previously we only kept a filename reference, which was actually used nowhere in the code yet. I replaced the simple list of filenames for a list of ngamsFileInfo objects, which can then be used for several purposes, such as serializing the object to XML, and more generally to store all the information about the files contained within a container