Improved container file handling / size calculation
Files are uniquely identified in the system by diskId, fileId (the basename of the file) and fileVersion. Two files with the same fileId but different fileVersions represent two different versions of the same file that have been stored. Two files with the same fileId and fileVersion, but different diskId, represent a different copy of the same file, which should have the same contents. After discussing with Andreas, we decided that containers should group files at the fileId level only, meaning that regardless where they are sitting and which version they have they should be considered as part of the container. In more practical terms this means that when retrieving a container the latest version of fileId will be retrieved as part of the container. To ensure this, we must pay special attention to the archiving command when submitting a new version of the file, as the containerId of the previous version must be passed onto the new version of the file. Also, when CAPPEND-ing or CREMOVE-ing files to/from the container it is safe (and easier) to simply update all rows for the given fileId. Finally, when CRETRIEVE-ing and CLIST-ing the latest version of each file must be retrieved.

On a separate note, the size of the container is now re-calculated when CAPPEND and CREMOVE are called, and when a new version of a file is submitted via QARCHIVE.