CAPPEND/CREMOVE work fine when skipping files
CAPPEND and CREMOVE update the size of the container using the uncompressed size of the file being added/removed to/from the container. If for some reason the file is actually not going through the intended operation the container size shouldn't be changed, which is what this patch solves. Also, in the case of multiple files being specified for both commands we were not actually performing the container size update.