Speed up own hostname lookups
The method that returns the name of the host to which the NGAS process first reads the Server/@IpAddress attribute from the configuration file. If this is not specified, it tries to figure out the machine's IP address using an external service on the internet. The result is then saved on a global variable to avoid further lookups. The problem was that further lookups where still being performed on every call to the method.

It should also be studied if we really want to contact an external service to give us our IP (which might be actually not ours, but a proxy's) instead of finding it out ourselves from the system hostname.