Proper handling of containerId in Janitor Thread
The Janitor Thread is reading the pickle files generated at ngamsFileInfo.write() time (issued during CARCHIVE and QARCHIVE for each individual file). In the previous version of the write method, the containerId was being saved along with the rest of the file information into the ngas_file table, and then into the pickle file, which then is read by the Janitor Thread and dumped once again into the database. This was causing race conditions when issuing QARCHIVE and CAPPEND commands in a quick fasion, because although the CAPPEND command was setting the containerId column to a value, the Janitor Thread could rewrite the row with the data from the QARCHIVE (i.e., containerId = null) AFTER the CAPPEND command finished, effectively taking the file out from the container.

To solve the issue, any writing to the containerId column is done outside the scope of the ngamsFileInfo.write() method. This way neither the original file writer command, neither the janitor thread, can write into the column.