Better integration of ingestion rate property
Since the beginning of the container branch a new "ingestion_rate" column was added to the ngas_files table to store the amount of time it took to ingest a particular file while archiving it. The column was there, but it wasn't being used consistently. Now it is exposed through the ngamsFileInfo class, through its XML representation, and is actually being filled with data during the QARCHIVE and CARCHIVE commands. The CARCHIVE is not calculating the correct numbers yet for each individual files though, and is storing the total time used to ingest the container instead.