Support for containers in postgreSQL DBs
A migration script has been written that can be run over existing postgreSQL databases to add the new columns/tables that are required to support containers in NGAS (plus the ingestion_rate column which was also added in the container branch).

On a different topic, the ngamsPostgreSQL driver wasn't properly handling exceptions, because it failed to raise them when it should have. This meant that the code was happily running when actually INSERTs and UPDATEs were not working at all. I corrected this error, and a different one that appeared after the exception handling was corrected, this time because the connect() method was not being called correctly.