Improving imports on pcc module
The pcc module now consists on a distribution that contains the pccDid, pccFits, pccKey, pccLog, pccPaf and pccUt packages. The __init__.py files of the packages don't handle anymore the __path__ or sys.path variables, so there is no automatic magical expansion of the python path at runtime. Also, the top-level pcc directory now *doesn't* contain an __init__.py file (which also edited the python path at runtime to include all other sub-packages directly into the path). With these changes it's not fully clear that the packages included in the pcc must be referenced in full name when being imported. Only the pcc/ top-level directory must be on the python path, and it remains to be the only directory that is on the python path after loading any of the pcc sub-packages.