Initial cleanup of ngas modules/packages
As before for the pcc package, now there is no more magic auto-expansion of the python path at runtime when the ngams package gets imported and its __init__.py file is processed. This file also exposed a number of methods and global "import" statements that made it difficult to track dependencies.

To fix both problems in one go, the __init__.py file has been deleted, and its contents (only the method and variable definitions, not the python path manipulation) has been moved to a new ngamsCore.py module under the ngamsLib/ package. Since *most* of the modules present on the ngams source tree used the magic provided by the ngams/ module by doing "from ngams import *", all these have had to be changed to import the necessary names from the new ngamsLib.ngamsCore module, meaning that a lot of files have been touched by this commit, although the changes themselves are quite small.

Now that there is no automatic inclusion of the ngams* packages into the python path at runtime, we either need to manually add the packages to the PYTHONPATH at runtime and continue referencing modules directly without their package names (e.g., "import ngamsCore"), or leave the PYTHONPATH set only to the level of the src/ directory and then reference all names using the package name (e.g., "from ngamsLib import ngamsCore") when importing modules from a different package. For the sake of clarity, and since I was editing a big number of files already, I decided to go with the latter option, which makes the setup of NGAS much easier, and the dependencies in the code much more clear.

Up to this commit the client and the server were successfully running (although not fully tested at all) when the PYTHONPATH was set to $PWD and $PWD/pcc (with PWD defined to ./ngas/src within this repository)