ngamsTests running
ngamsTests now run, or at least seem to, because they have been ported to the new way of importing modules. A small importing problem was also found in pcc and fixed.

The main remaining problem to fix is to solve the name clashes. ngamsLib, ngamsPClient, ngamsServer and ngamsTest reference at the same time a package and a module. This means that when inside one of these packages, modules cannot reference other modules within the same package using their fully-qualified name (i.e., package.module, which is what Eclipse does automatically) because the package name is interpreted as the module name. This is something we can live with probably, but is confusing and can probably be improved.