Solving circular package dependency
The pccUt package sits at the bottom of the pcc* packages in terms of dependency. The PccUtDb2KeyMap module was the only exception to this rule, which was pulling modules from other packages, which in turn depend on pccUt. Moving this sole module to a new package solves the problem.