Centralizing LICENSE, VERSION and COPYRIGHT files
The LICENSE, VERSION and COPYRIGHT files were duplicated in different directories, some even having different contents. The files now live exclusively in the root directory of the ngas project. Symbolic links have been added from the ngamsData/ package so these files end up included in the ngamsCore egg, from where they are exposed to the rest of the code via the different mechanisms available in ngamsLib.ngamsCore. With this change the ngamsPClient finally runs successfully, and probably the ngamsServer as well