Improving plugIn loading
Loading NGAMS PlugIns has so far been always a manual task including exec and eval instructions that are error prone. Instead we now use the importlib module to import the module (using the fully-qualified name, i.e. ngamsPlugIns.plugInModuleName), and then we get the plugin's entry point method reference via getattr(). This logic is centralized in a new method under the ngamsCore module which loads both the module and the entry point method (optionally with a different name from the plugIn module) and returns the method for direct usage.