Big cleanup towards removal of buildout
Buildout seems like a very big solution to a problem that doesn't exist. It pulls out a big number of eggs from the internet (or locally, from the additional_tars/eggs.tar.gz tarball), installs a big number of dependencies, and all to finally perform a typical ./configure && make && make install. It also doesn't actually build binary eggs and install them in some given location, but instead it simply creates "development" eggs and leaves files as they are and where they are. The buildout (stubborn, as they say themselves) philosophy is that projects should be self-contained, but on the other hand we're using virtualenv to set up a separate virtual environment. I'm trying here to move away from buildout and towards virtualenv with eggs installed on it.