PccUtUtils.execCmd improved
Processes that finished too early (i.e., before the first call to poll()) didn't report any stdout or stderr messages back to the caller. This was affecting tests that quickly failed but didn't report a single line about what the process actually outputed.

Since I was changing the code, I took the change to upgrade it to use subprocess instead of the old popen2 package