Making tests more runnable
The main change is that the ngamsFitsPlugin now optionally can skip checksum validation, which requires a custom binary to be present on the PATH (and which cannot be found anywhere on the internet, seems to have been an ESO-specific program). The other changes are to start testing servers on all IPs and to use a no-op label plugin (for which a new no-op plugin has been added).