First take on modularized fabfile
Instead of maintaining a huge fabfile module that contains dozens and dozens of methods that are widly and obscurely connected we now have a fabfile package with separate modules for separate groups of tasks and methods. The code has been cleaned up to be more clear, have less interdependencies, and ultimately to make sure it carries out the installation of NGAS in a variety of use cases.