ngamsServer exiting and ngamsDaemon sanitized
The way in which the ngamsServer was being terminated before wasn't nice, since it always finished with a suicide (i.e., a os.kill(SIGKILL) on itself). This could be useful during testing but not on an operational system. There was also a bit of repetitive code which could be brought together. Finally, the http server itself (i.e., the HTTPBaseServer instance) wasn't being shutdown (which is probably why they were SIGKILL-ing the process).

Now things look way better: no SIGKILL is sent at any time, and the ngamsServer termination process is well centralized, and invoked from several places using a simple call (including the SIGTERM and SIGINT signal handlers, and the EXIT command handler). The termination include the proper shutdown of the HTTP server, plus the original clean-up actions that before were occuring before the SIGKILL was being sent.

To wrap all up, I re-wrote most of the ngamsDaemon module to properly start and end the ngamsServer using the PIP-standard daemon facilities instead of our own home-brewed daemon class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>