Getting rid of global getHostId() et al.
This is probably the main commit on this branch. It completely removes the process-global getIpAddress and getHostId calls and puts them into the ngamsServer class, since that's where they belonged in the first place. Lots of code everwhere used this global methods (although some of them only for tiny things and didn't really needed it) which is why the commit is so big and touches so many files.

Along with this big refactoring a few small cleanups are included, like removing unused code and such.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>