Getting rid of "reference config" on unit tests
The reference configuration is something that wasn't necessary and therefore made things more complex than they should have been. The only reason I had left it in place so far is because the old global getHostName() and getHostId() implementations relied on a '-cfg' parameter given on the command-line, which these tests did. Now that this dependency is not there anymore we can stop using a -cfg during the tests to specify a reference configuration that wasn't really needed.

All in all, it's a win-win situation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>