Removing ngamsLib._setSocketTimeout
This method was either setting a timeout *globally* as the default timeout for
all sockets, or only on the socket underlying a given httplib.HTTP object.
Moreover, if a None timeout was given it would default to a 3600 (i.e., AN
HOUR) timeout. It was of course being abused, and a global timeout was being
set in situations where a local timeout was meant (which are normally almost
ALL situations one could think of).

Removing this method gives me a warm feeling.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>