Server-as-client timeouts improved
The NGAS server acts sometimes as a client (e.g., when fetching data from
another server to hand it over to a client later). In such cases the original
client can specify an optional time_out parameter to control how long the
server should wait when dealing with the internal one. As implemented before
this commit the parameter was actually mandatory, and allowed the special -1
value to signify "no timeout" situations. This has been changed now so the
parameter is optional, in which case a hardcoded default timeout applies.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>