Avoid clashes on tmp filenames
If more than one server was started within the same second we were experiencing
a race condition where the first server was pointing to the configuration of
the second. Using more unique names for these files solves the issue, and we
should probably look at using the tempfile package for these things anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>