Removing unnecessary pcc modules
This is a big cleanup of the pcc packages. Most of them were not being used at
all by the NGAS code (directly or indirectly, I used snakefood to track down
the dependencies) so removing them was trivial. In a handful of cases the code
was easily replaceable with standard python code which is why there are small
changes on the NGAS code itself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>