Don't TRACE() during __del__
If python is deleting modules (provoking a call to __del__ methods) some global
names might not be available already during the call to TRACE(), so an
exception will be thrown during __del__ which finally shows itself as a warning
message in stderr. This is why it's discouraged to perform anything outside the
strictly fundamental and to maintain external invariants.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>