Improved test server shutdown
A timeout was needed; otherwise the OFFLINE command would hang forever. I've
also removed the use of getHostName() when creating clients on the tests to
avoid connecting through anything that isn't the loopback interface. This is no
so critical in normal scenarios, but when testing in more restricted
environments (like in AWS, where there are strict rules about which ports are
exposed through the public network interface) this quickly becomes a problem.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>