Revert to IDLE when not really BUSY with ARCHIVE
The ARCHIVE command starts always by setting the server as BUSY, but wasn't
setting the it back to IDLE when returning quickly (e.g., when probing it or
when sending back a redirect response). This affects of course any other
command that relies on the server being in a certain substate. Two notable
cases are the OFFLINE (without force) and EXIT commands, which require the
server to be in the IDLE state.

Some unit tests were suffering because of this, in particular those that were
doing archiving via a frontal proxy server which was then redirecting the
client to the final node that does the archiving. The unit tests shut down the
servers that are brought up during the tests by sending them an OFFLINE and an
EXIT command, or resorting to SIGKILL them if the above times out. We were
continuously timing out in the described scenarios, and therefore the unit
tests took ages to run in some cases.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>