First iteration of PEP249-ification of NGAS
This change is all about stop using the NGAS-custom DB plugins and start using
the PE249-compliant DB modules that each vendor distributes. The concepts are
strikingly similar because the NGAS DB plugin API was built around these PEP249
modules (with or without intention), so only the core class that provides DB
access needed to be changed while all its users remain for the time being
completely untouched.

There are still a few things that need to be cleaned up and clarified, but at
least the current version already works fine for a manually-started,
automatic-online NGAS server and in the ngamsContainer unit tests (haven't
tried them all yet mainly because they take so long, but will surely run them
in the future -- having at least one running fine makes me feel already certain
that most of the stuff is working already anyway). At the same time I have
restricted the testing to SQlite 3 only, so running with a different database
in the future is also a must.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>