Cleaned up ngamsDb hierarchy
On the one hand there were a number of unnecessary constructors that were doing
nothing but passing down the same parameters with the same names and same
defaults down to the parent constructor, so they were useless. On the other
hand the ngamsDbBase class made no sense in the hierarchy so I completely
removed it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>