New ngamsDb construction
The ngamsDb object is now constructed with fewer parameters than before. The
class structure previously included things like username, password and a
server, which do not apply to all database types. On the other hand we needed a
design that would be flexible enough to allow us to pass any information down
to the connection logic.

This new mechanism is implemented by allowing extra attributes to be present
under the Db element of the NGAS server XML configuration. These attributes'
values are casted into booleans, ints, floats or simply string as required, and
passed down to the ngamsDb object, which internally uses them to open the
connection to the database.

I took the opportunity thus to centralize the creation of ngamsDb objects using
a ngamsConfig object. Several lines of code have been trimmed out from both the
server and the tests. At the same time I cleaned up a bit the base test class
to improve how the configuration for the external servers was being calculated.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>