New method for parametric queries
A new query2 method is being added in this commit, plus examples on how to use
it, to perform parametric queries instead of plain queries onto the database.
query2 is also substantially simpler than its older query sibling as it doesn't
have automatic retry and reconnection capabilities, leaving that concern to the
user instead. Finally, the return type of this method is a simpler 2D structure
(or an empty array in the case of no results), which is simpler than the 3D
structure returned by query.

query2 accepts queries both with or without parameters. A parametrized query
must use the {X} notation for each placeholder; this way we can internally
populate the string with the appropriate parameter placeholder depending on the
database module's supported paramstyle as indicated in PEP-249. The same way we
internally sanitize the argument list into whatever required format is needed
by the cursor.execute method of the database module.

Adding a new method instead of simply replacing the old one makes it easier to
transition from one to the other without breaking the whole software.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>