Ported ngamsDbNgasCache to query with parameters
Some methods of this class were not called at all, so I also took the
opportunity to remove them. It also seems that all the cache-related stuff that
happens on the ngamsCacheControlThread module is using its own sqlite database,
either *instead of* or *as well as* the main database. We would have to look
into this more closely, but it's also outside the scope of the current
exercise.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>