Bulk porting from .query to .query2
Changes are mostly on the ngamsPlugIns package. A few queries that repeated
themselves across several plug-ins were moved to the ngamsDb* classes, and the
rest have simply been ported to .query2.

This finishes the main transition to parametrized queries in the NGAS code. The
only remaining bits are those under ngasUtils, which can be ported bit by bit
when needed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>