Initial version of proper timeout handling
The previous code was based on continuously invoking poll() on *blocking*
sockets, which is a bit of a nonesense. Given that the I/O operations on
blocking sockets respect the socket timeouts we are now setting the time out on
the socket descriptor instead, and properly checking the return values of the
I/O operations.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>