Unifiying checksum-related logic
The checksum-related logic is now unified under the ngamsFileUtils module. This
new centralized checksum logic takes into account both numeric and string CRC
variant names/identifiers, and performs file checksum checking in a central
place. It also returns the corresponding crc method given a variant specifier.

A new test has been written to double-check that this is working as expected.

There are still other checksum checks performed by plug-ins (namely the
ngamsGenDapiPlugIn and others that have been copied from it), where a
checkChecksum method runs a user-supplied command to verify a user-supplied
checksum value -- in other words, this is completely isolated from the main
checksum calculation. Since it adds an extra checksum step with a potentially
blocking command this should be removed, and replaced with a simple comparison
between the main checksum calculated during archiving and the checksum given by
the user.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>