Support both ngas_files.ignore/file_ignore columns
From what we can re-construct now it seems that the first version of the
ngas_files table had an 'ignore' column. The word 'ignore' was then found to be
an unacceptable solution because some database engines have an 'ignore' keyword
which cannot be used as a column name, and thus the column was renamed to
'file_ignore'. This transition was not properly carried out in the code, so
there was a mixture of both column names being used here and there, with
'file_ignore' dominating most of the scene.

This problem was hidden by the old code because the old ngas-specific DB
plug-ins were doing a very rough string replacement on the SQL query before
submitting it to the actual python driver module, changing from one column name
to the other. This effectively avoided the problem but had the potential of
introducing others even more difficult to spot, as actual data could have been
changed during the string replacement. During unit testing the problem didn't
show up, as most of the code uses the 'file_ignore' and the database set up for
unit tests uses the same column name, and therefore after the redesign of the
DB code in ngas some months ago the problem still didn't show up.

Finally, the problem became finally evident while testing an NGAS installation
against a production PostgreSQL database. This database was using the 'ignore'
column, and thus the code broke everywhere. Because this database is still used
by NGAS servers based on the old code we have to be careful and properly
support both of them at the same time.

We came up with a temporal solution for this problem, which is flexible and
correct. We allow users to choose which columnname should be used via a new
UseFileIgnore attribute on the Db element of the XML server configuration file,
which defaults to True. The code underneath then uses either 'ignore' or
'file_ignore' accordingly. We expect to use this feature while rolling out the
new server in production environments that need to still support the old server
as well. Once all servers have been replaced with this newer version the
database can be safely ported to use 'file_ignore' and we can remove this
option again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>