crc variant == -1 means that no CRC is performed
This has been the case in some production systems because the burden imposed by
the CRC was too big. Therefore it's reasonable to expect the same limitations
in other systems.

In the cases when no CRC is calculated, a NULL is stored in both the checksum
and the checksum_plugin columns.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>