Fixes on apt-get package management
First, each package will need to be installed in a different apt-get install
invocation. This is because if any of the package names doesn't exist for some
reason the whole install will be aborted, and the rest of the packages will not
be installed.

The zlib1g-dgb package contains only the stripped debugging symbols from the
runtime library, while the -dev package contains the development headers, etc,
which are actually required by the C client compilation.

The openssh-server package installation doesn't make sense, as the whole fab
script is run *through* SSH in the first place.

Finally, the python-setuptools seems to be not always installed, at least in
Ubuntu 12.04, but it's required to create a virtualenv.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>