An improved version of check_ssh
This new version makes sure that a big, clear message is displayed when no SSH
connection can be made to the target host. If the target host is localhost then
an extra instruction is given on how to set up the SSH server in the local
machine.

This method is now called at the *very* beginning of the hl.user_deploy and
hl.operations_deploy tasks. It also required a small fix on the is_localhost
method to ensure it doesn't unnecessarily execute something if we already have
a host in our environment (such execution would fail if we have a target host
in the env alread, but it doesn't have SSH running, which is exactly when we
are calling the function from check_ssh).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>