Check which python during virtualenv_setup
The main reason to do this is that the user might want to overwrite a previous
installation. In this case the python binary to be used for creating the
virtualenv would have been that inside the virtualenv, which is not available
after removing the virtualenv of course. We thus now first (optionally, as
always) remove the virtualenv and then calculate which python to use for
creating the new virtualenv.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>