ngamsTest running without need to install them
The ngamsTest package is actually not meant to be ever installed (and it's not
installed by build.sh actually). On the other hand some tests require to load
plug-ins that are defined in the ngamsTests package, and therefore they still
required the package to be installed, or be in the PYTHONPATH for that matter.

With this change we manually set the PYTHONPATH of the server subprocesses so
it includes the ngamsTest package on its namespace, therefore allowing it to
load those plugins.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>