Correctly handle fab's env.key_filename
Its value can be either a string or a list, so we have to handle both fine.

Also, now if a key is given on env.key_filename only that is used without
trying ~/.ssh/id_rsa later (which will be used only when no env.key_filename is
present).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>