Fixing imports in newest GLEAMCUTOUT
Some names that before were imported via ngams.* are not there anymore. This
had been fixed already in the past, but in the last rewriting of the module
those changes were lost.

I also had to change some tiny bits of logic as importing ngams.* will always
yield an ImportError. All the branches that relied on it working are thus false
and were pruned.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>