New NGAS_DEVELOP variable for fabric
This new option allows users to specify that the NGAS Python packages installed
via the fabric process should be installed in development mode. This flag
internally invokes the build.sh script with the -d option, which in turn calls
"python setup.py develop" on each of the python packages (as opposed to "python
setup.py install").

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>