Improved user/operational_deploy logic separation
I've rearranged the code of the main task logic to be more readable than
before. The operational_deploy (and the tasks using it) were always setting the
install() task parameters to True, while the user_deploy was setting them all
to False. Instead of having three different parameters I thus decided to have a
separate method for the operational_deploy which would proceed with the
system-level setup, and then call the usual user_deploy steps.

While doing this it was also clear that the NGAS_USER concept only makes sense
when doing an operational_deploy (and friends); the user_deploy on the other
hand should use the ssh user for installing NGAS. I thus changed the code to
act accordingly, which made it (in my opinion) even more readable than before.
The documentation has been changed accordingly as well.

Some other cleanups made it to this commit also.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>