Task descriptions improved.
The full descriptions of the tasks we had in the install.rst file have now been
separated into their own .rst file, and are appended to the short __doc__ of
each of the corresponding hl.* task. This way users can do fab -d <task> and
get comprehensive description of each of the high-level tasks, together with
all the fabric variables that affect the behavior of the tasks.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>