Getting rid of pccFits
The only use of pccFits in the ngams code (actually only in the ngamsPlugIns
package) was to either check if a header existed in a fits file, or to actually
read those headers (from the primary HDU). On the other hand we are already
using pyfits in other plugins to perform fits-related operations, and therefore
it made no sense to continue using our own home-brewed fits parser (which had a
few shortcomings anyway).

With this change, and together with 2f7664e, we really are using pyfits
everywhere instead of two or three different solutions.

This feels good.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>