Getting rid of PccUtUtils
The only meaningful method there was execCmd. This was already being thinly
wrapped by the ngamsPlugInApi module and offered to the plug-ins that way.
There were only a few other places where this was used, and therefore I moved
it down to ngamsCore.

Having said this, there is a troubling number of modules that implement their
own execCmd function, most of them looking very similar, so a proper cleanup of
the execCmd situation is now in the TODO list.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>