Removing unnecessary code from PccUtTime
Changed include removing a few methods nobody used, unnecessary type checks and
arguments in some of the functions, and unnecessary special handling of the
syslog module, among others.

The ultimate goal is to get rid of it, but the Timestamp and Timer classes are
used in too many places, and appear to affect a lot the logic of the
application. Excessive care will have to be exercised when removing those
classes, making sure that the timestamps are still correctly stored in the
database, and read back correctly, printed correctly, etc.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>