Removing ngamsCore.logSetCond
This method is not universal in the sense that the logic it contains doesn't
apply to all the programs that want to use NGAS libraries (e.g., it setups up a
syslog handler and a file handler, both of which are usually not needed by
normal applications. The body of this function now lives in the ngamsServer
class under the setup_logging method, and has been reworked to setup the
standard logging library instead of using the PccLog classes.

Usage of the old logSetCond has been erased throughout the codebase (except
from the ngasUtils scripts), and each main program is thus left to configure
its own logging in whatever way they see fit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>