Adjusting tests that relied on particular logs
A few of the tests relied on the value of particular log lines, which were
manually parsed to extract information for comparisson. This is of course not a
nice practice, but on the other hand properly re-implementing them would take
quite some effort. What we can do in the meantime is to fix them so they
continue passing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>