New from/to iso8601/mjd functions
The from/to iso8601 functions are meant to replace the several existing string
<--> time functions throughout the code, primarily those under PccUtTime (with
the purpose of finally removing that module), and unify them. They work both
with UTC (default) and local time conversions and offer different formatting
options, and thus should accommodate for most of the current use cases.

The from/to mjd functions are meant to replace the functionality found inside
the PccUtTime class at the moment, but without the need of constructing a class
around the calculation, thus making it more straightforward.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>