Removing timeRef2Iso8601
This function was used all throughout the code, specially to sanitize inputs
for several attributes of the domain classes used by NGAS, which internally
stored timestamps as local ISO-8601 strings. Now they all (hopefully! at least
all tests pass) internally store the number of seconds since the epoch, some of
them are None-able, and we are nicely converting from and to the string
representation as required.

This is a big change, and I really hope nothing is breaking due to this. The
fact that the tests are working gives me some confidence already, but some dark
corners might have left untouched and thus broken.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>