Unified interface of janitor plugins
Code that was needed by some of them now resides in a common module which both
the ngamsJanitorThread module and the plugin modules can import and use.

With a common interface now the process of loading janitor plugin modules and
running through them on each cycle is greatly simplified. A hardcoded list of
plugins is currently maintained, but in the future users should be able to add
more plugins in the configuration file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>