Cleaning up ngamsDbm module and usage
The ngamsDbm2 class was not used (and thus the dbm module doesn't need to be
imported anymore). To be more precise, it *seems* to be used by the
ngasXSyncTool script, but it can be safely replaced by a normal ngamsDbm class
(I didn't do it though as the script is probably not being used at all).

The other main change is the translation from a bsddb-specific
DbRunRecoveryError exception into our ngams-specific one. This allows us to
handle this specific exception in the code without tricks like "if
str(e).find('this_or_that')" while maintaining the bsddb-specifics within the
ngamsDbm module only.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>