Removing the "wait" flag
The "wait" flag was present across all NGAS commands (e.g., it was explicitly
listed in the ngamsReqProps class which represents a generic set of request
properties). Its value though was used only by a few commands, namely OFFLINE,
ARCHIVE, REGISTER and CLONE. It made sense then to expose this flag only for
those commands, and not treat it as a generic flag that any command could use.

The name "wait" was also confusing because it wasn't clear who was doing the
waiting, and therefore I've renamed it to "async" which refers to the operation
as a whole. Therefore what previously was wait=0 now is async=1 and vice-versa.

Also, to aid the inspection of values in the ngamsReqProps class I've
implemented the python container interface methods on it so it's easier now to
see if a given property is present in the object, and to retrieve its value if
necessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>