Removed commented RETRIEVE code and tests
The RETRIEVE command allowed in the past to retrieve the NGAS configuration
file, the NGAS log file, and any internal NGAS file. This was commented out
some time ago (because it was obviously a security risk) together with the
tests that exercised that functionality. Now that time has passed I'm simply
deleting the commented code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>